+++ tests/examples/helm-local/index.ts
@@ -21,6 +21,7 @@ const namespaceName = namespace.metadata.name;
 const nginx = new k8s.helm.v2.Chart("simple-nginx-local", {
     // Represents chart `stable/nginx-lego@v0.3.1`.
     path: "nginx-lego",
+    namespace: namespaceName,
     values: {
         // Override for the Chart's `values.yml` file. Use `null` to zero out resource requests so it
         // can be scheduled on our (wimpy) CI cluster. (Setting these values to `null` is the "normal"
@@ -39,14 +40,6 @@ const nginx = new k8s.helm.v2.Chart("simple-nginx-local", {
                 }
             }
         },
-        // Put every resource in the created namespace.
-        (obj: any) => {
-            if (obj.metadata !== undefined) {
-                obj.metadata.namespace = namespaceName
-            } else {
-                obj.metadata = {namespace: namespaceName}
-            }
-        }
     ]
 });